+++ example/views.py
@@ -5,9 +5,11 @@ import rest_framework_json_api.metadata
 import rest_framework_json_api.parsers
 import rest_framework_json_api.renderers
 from rest_framework_json_api.views import ModelViewSet, RelationshipView
-from example.models import Blog, Entry, Author, Comment
+from example.models import Blog, Entry, Author, Comment, Company, Project
 from example.serializers import (
-    BlogSerializer, EntrySerializer, AuthorSerializer, CommentSerializer)
+    BlogSerializer, EntrySerializer, AuthorSerializer, CommentSerializer, CompanySerializer,
+    ProjectSerializer,
+)
 
 from rest_framework_json_api.utils import format_drf_errors
 
@@ -69,6 +71,16 @@ class AuthorViewSet(ModelViewSet):
 class CommentViewSet(ModelViewSet):
     queryset = Comment.objects.all()
     serializer_class = CommentSerializer
+
+
+class CompanyViewset(ModelViewSet):
+    queryset = Company.objects.all()
+    serializer_class = CompanySerializer
+
+
+class ProjectViewset(ModelViewSet):
+    queryset = Project.objects.all()
+    serializer_class = ProjectSerializer
 
 
 class EntryRelationshipView(RelationshipView):