+++ rest_framework_json_api/relations.py
@@ -21,6 +21,7 @@ LINKS_PARAMS = [
 
 
 class ResourceRelatedField(PrimaryKeyRelatedField):
+    _skip_polymorphic_optimization = True
     self_link_view_name = None
     related_link_view_name = None
     related_link_lookup_field = 'pk'
@@ -168,7 +169,7 @@ class ResourceRelatedField(PrimaryKeyRelatedField):
             pk = value.pk
 
         resource_type = self.get_resource_type_from_included_serializer()
-        if resource_type is None:
+        if resource_type is None or not self._skip_polymorphic_optimization:
             resource_type = get_resource_type_from_instance(value)
 
         return OrderedDict([('type', resource_type), ('id', str(pk))])
@@ -222,6 +223,48 @@ class ResourceRelatedField(PrimaryKeyRelatedField):
             )
             for item in queryset
         ])
+
+
+class PolymorphicResourceRelatedField(ResourceRelatedField):
+    """
+    Inform DRF that the relation must be considered polymorphic.
+    Takes a `polymorphic_serializer` as the first positional argument to
+    retrieve then validate the accepted types set.
+    """
+
+    _skip_polymorphic_optimization = False
+    default_error_messages = dict(ResourceRelatedField.default_error_messages, **{
+        'incorrect_relation_type': _('Incorrect relation type. Expected one of [{relation_type}], '
+                                     'received {received_type}.'),
+    })
+
+    def __init__(self, polymorphic_serializer, *args, **kwargs):
+        self.polymorphic_serializer = polymorphic_serializer
+        super(PolymorphicResourceRelatedField, self).__init__(*args, **kwargs)
+
+    def to_internal_value(self, data):
+        if isinstance(data, six.text_type):
+            try:
+                data = json.loads(data)
+            except ValueError:
+                # show a useful error if they send a `pk` instead of resource object
+                self.fail('incorrect_type', data_type=type(data).__name__)
+        if not isinstance(data, dict):
+            self.fail('incorrect_type', data_type=type(data).__name__)
+
+        if 'type' not in data:
+            self.fail('missing_type')
+
+        if 'id' not in data:
+            self.fail('missing_id')
+
+        expected_relation_types = self.polymorphic_serializer.get_polymorphic_types()
+
+        if data['type'] not in expected_relation_types:
+            self.conflict('incorrect_relation_type', relation_type=", ".join(
+                expected_relation_types), received_type=data['type'])
+
+        return super(ResourceRelatedField, self).to_internal_value(data['id'])
 
 
 class SerializerMethodResourceRelatedField(ResourceRelatedField):