+++ rest_framework_json_api/renderers.py
@@ -12,7 +12,7 @@ from rest_framework import renderers
 from rest_framework.serializers import BaseSerializer, Serializer, ListSerializer
 from rest_framework.settings import api_settings
 
-from . import utils
+from rest_framework_json_api import utils
 
 
 class JSONRenderer(renderers.JSONRenderer):
@@ -344,8 +344,6 @@ class JSONRenderer(renderers.JSONRenderer):
                 relation_type = utils.get_resource_type_from_serializer(serializer)
                 relation_queryset = list(relation_instance)
 
-                # Get the serializer fields
-                serializer_fields = utils.get_serializer_fields(serializer)
                 if serializer_data:
                     for position in range(len(serializer_data)):
                         serializer_resource = serializer_data[position]
@@ -354,12 +352,18 @@ class JSONRenderer(renderers.JSONRenderer):
                             relation_type or
                             utils.get_resource_type_from_instance(nested_resource_instance)
                         )
+                        serializer_fields = utils.get_serializer_fields(
+                            serializer.__class__(
+                                nested_resource_instance, context=serializer.context
+                            )
+                        )
                         included_data.append(
                             cls.build_json_resource_obj(
                                 serializer_fields,
                                 serializer_resource,
                                 nested_resource_instance,
-                                resource_type
+                                resource_type,
+                                getattr(serializer, '_poly_force_type_resolution', False)
                             )
                         )
                         included_data.extend(
@@ -381,7 +385,8 @@ class JSONRenderer(renderers.JSONRenderer):
                     included_data.append(
                         cls.build_json_resource_obj(
                             serializer_fields, serializer_data,
-                            relation_instance, relation_type)
+                            relation_instance, relation_type,
+                            getattr(field, '_poly_force_type_resolution', False))
                     )
                     included_data.extend(
                         cls.extract_included(
@@ -423,7 +428,11 @@ class JSONRenderer(renderers.JSONRenderer):
         return data
 
     @classmethod
-    def build_json_resource_obj(cls, fields, resource, resource_instance, resource_name):
+    def build_json_resource_obj(cls, fields, resource, resource_instance, resource_name,
+                                force_type_resolution=False):
+        # Determine type from the instance if the underlying model is polymorphic
+        if force_type_resolution:
+            resource_name = utils.get_resource_type_from_instance(resource_instance)
         resource_data = [
             ('type', resource_name),
             ('id', encoding.force_text(resource_instance.pk) if resource_instance else None),
@@ -506,6 +515,9 @@ class JSONRenderer(renderers.JSONRenderer):
             # Get the serializer fields
             fields = utils.get_serializer_fields(serializer)
 
+            # Determine if resource name must be resolved on each instance (polymorphic serializer)
+            force_type_resolution = getattr(serializer, '_poly_force_type_resolution', False)
+
             # Extract root meta for any type of serializer
             json_api_meta.update(self.extract_root_meta(serializer, serializer_data))
 
@@ -517,7 +529,7 @@ class JSONRenderer(renderers.JSONRenderer):
                     resource_instance = serializer.instance[position]  # Get current instance
 
                     json_resource_obj = self.build_json_resource_obj(
-                        fields, resource, resource_instance, resource_name
+                        fields, resource, resource_instance, resource_name, force_type_resolution
                     )
                     meta = self.extract_meta(serializer, resource)
                     if meta:
@@ -532,7 +544,7 @@ class JSONRenderer(renderers.JSONRenderer):
             else:
                 resource_instance = serializer.instance
                 json_api_data = self.build_json_resource_obj(
-                    fields, serializer_data, resource_instance, resource_name
+                    fields, serializer_data, resource_instance, resource_name, force_type_resolution
                 )
 
                 meta = self.extract_meta(serializer, serializer_data)